Move examples into the connector
Now we can use "cargo run --example point_query" to run this example. Dependencies are automatically the same as for the library.
