--- examples/netcdf_queries/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "netcdf_queries"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-rust-connector-api = { path = "../../meteomatics"}
-tokio = { version = "1", features = ["full"] }
-chrono = "0.4"
-polars = "0.21.1"
-netcdf = "0.7.0"
-dotenv = "0.15.0"
