--- examples/point_queries/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "point_queries"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-rust-connector-api = { path = "../../meteomatics"}
-tokio = { version = "1", features = ["full"] }
-chrono = "0.4"
-polars = "0.21.1"
-dotenv = "0.15.0"
